SD-2258: Refactor class names
