+++ booking/src/main/java/org/dcsa/conformance/standards/booking/action/UC11ShipperCancelBookingRequestAction.java
@@ -1,6 +1,7 @@
 package org.dcsa.conformance.standards.booking.action;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import java.util.stream.Stream;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.dcsa.conformance.core.check.*;
@@ -8,17 +9,15 @@ import org.dcsa.conformance.core.traffic.HttpMessageType;
 import org.dcsa.conformance.standards.booking.party.BookingRole;
 import org.dcsa.conformance.standards.booking.party.BookingState;
 
-import java.util.stream.Stream;
-
 @Getter
 @Slf4j
-public class UC11_Shipper_CancelBookingRequestAction extends StateChangingBookingAction {
+public class UC11ShipperCancelBookingRequestAction extends StateChangingBookingAction {
   private final JsonSchemaValidator requestSchemaValidator;
   private final JsonSchemaValidator responseSchemaValidator;
   private final JsonSchemaValidator notificationSchemaValidator;
   private final BookingState expectedBookingStatus;
 
-  public UC11_Shipper_CancelBookingRequestAction(
+  public UC11ShipperCancelBookingRequestAction(
       String carrierPartyName,
       String shipperPartyName,
       BookingAction previousAction,
