+++ booking/src/main/java/org/dcsa/conformance/standards/booking/action/UC7ShipperSubmitBookingAmendment.java
@@ -1,6 +1,7 @@
 package org.dcsa.conformance.standards.booking.action;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import java.util.stream.Stream;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.dcsa.conformance.core.check.*;
@@ -8,27 +9,25 @@ import org.dcsa.conformance.core.traffic.HttpMessageType;
 import org.dcsa.conformance.standards.booking.party.BookingRole;
 import org.dcsa.conformance.standards.booking.party.BookingState;
 
-import java.util.stream.Stream;
-
 @Getter
 @Slf4j
-public class UC9_Shipper_CancelBookingAmendment extends StateChangingBookingAction {
+public class UC7ShipperSubmitBookingAmendment extends StateChangingBookingAction {
   private final JsonSchemaValidator requestSchemaValidator;
   private final JsonSchemaValidator responseSchemaValidator;
   private final JsonSchemaValidator notificationSchemaValidator;
   private final BookingState expectedBookingStatus;
   private final BookingState expectedAmendedBookingStatus;
 
-  public UC9_Shipper_CancelBookingAmendment(
-    String carrierPartyName,
-    String shipperPartyName,
-    BookingAction previousAction,
-    BookingState expectedBookingStatus,
-    BookingState expectedAmendedBookingStatus,
-    JsonSchemaValidator requestSchemaValidator,
-    JsonSchemaValidator responseSchemaValidator,
-    JsonSchemaValidator notificationSchemaValidator) {
-    super(shipperPartyName, carrierPartyName, previousAction, "UC9", 202);
+  public UC7ShipperSubmitBookingAmendment(
+      String carrierPartyName,
+      String shipperPartyName,
+      BookingAction previousAction,
+      BookingState expectedBookingStatus,
+      BookingState expectedAmendedBookingStatus,
+      JsonSchemaValidator requestSchemaValidator,
+      JsonSchemaValidator responseSchemaValidator,
+      JsonSchemaValidator notificationSchemaValidator) {
+    super(shipperPartyName, carrierPartyName, previousAction, "UC7", 202);
     this.requestSchemaValidator = requestSchemaValidator;
     this.responseSchemaValidator = responseSchemaValidator;
     this.notificationSchemaValidator = notificationSchemaValidator;
@@ -39,12 +38,7 @@ public class UC9_Shipper_CancelBookingAmendment extends StateChangingBookingActi
   @Override
   public String getHumanReadablePrompt() {
     return getMarkdownHumanReadablePrompt(
-        "prompt-shipper-uc9.md", "prompt-shipper-refresh-complete.md");
-  }
-
-  @Override
-  protected boolean expectsNotificationExchange() {
-    return true;
+        "prompt-shipper-uc7.md", "prompt-shipper-refresh-complete.md");
   }
 
   @Override
@@ -56,6 +50,11 @@ public class UC9_Shipper_CancelBookingAmendment extends StateChangingBookingActi
   }
 
   @Override
+  protected boolean expectsNotificationExchange() {
+    return true;
+  }
+
+  @Override
   public ConformanceCheck createCheck(String expectedApiVersion) {
     return new ConformanceCheck(getActionTitle()) {
       @Override
@@ -63,18 +62,18 @@ public class UC9_Shipper_CancelBookingAmendment extends StateChangingBookingActi
         var dsp = getDspSupplier().get();
         String reference = dsp.carrierBookingReference() !=  null ? dsp.carrierBookingReference() : dsp.carrierBookingRequestReference();
         return Stream.concat(
-          Stream.concat(createPrimarySubChecks("PATCH", expectedApiVersion, "/v2/bookings/%s".formatted(reference)),
+          Stream.concat( createPrimarySubChecks("PUT", expectedApiVersion, "/v2/bookings/%s".formatted(reference)),
             Stream.of(
               new JsonSchemaCheck(
                 BookingRole::isShipper,
                 getMatchedExchangeUuid(),
                 HttpMessageType.REQUEST,
                 requestSchemaValidator))),
-          getNotificationChecks(
-            expectedApiVersion,
-            notificationSchemaValidator,
-            expectedBookingStatus,
-            expectedAmendedBookingStatus));
+            getNotificationChecks(
+                expectedApiVersion,
+                notificationSchemaValidator,
+                expectedBookingStatus,
+                expectedAmendedBookingStatus));
       }
     };
   }
