+++ booking/src/main/java/org/dcsa/conformance/standards/booking/party/BookingShipper.java
@@ -58,17 +58,17 @@ public class BookingShipper extends ConformanceParty {
   @Override
   protected Map<Class<? extends ConformanceAction>, Consumer<JsonNode>> getActionPromptHandlers() {
     return Map.ofEntries(
-        Map.entry(UC1_Shipper_SubmitBookingRequestAction.class, this::sendBookingRequest),
+        Map.entry(UC1ShipperSubmitBookingRequestAction.class, this::sendBookingRequest),
         Map.entry(ShipperGetBookingAction.class, this::getBookingRequest),
-        Map.entry(Shipper_GetAmendedBooking404Action.class, this::getBookingRequest),
-        Map.entry(UC3_Shipper_SubmitUpdatedBookingRequestAction.class, this::sendUpdatedBooking),
-        Map.entry(UC7_Shipper_SubmitBookingAmendment.class, this::sendUpdatedConfirmedBooking),
-        Map.entry(UC9_Shipper_CancelBookingAmendment.class, this::sendCancelBookingAmendment),
-        Map.entry(UC11_Shipper_CancelBookingRequestAction.class, this::sendCancelBookingRequest),
+        Map.entry(ShipperGetAmendedBooking404Action.class, this::getBookingRequest),
+        Map.entry(UC3ShipperSubmitUpdatedBookingRequestAction.class, this::sendUpdatedBooking),
+        Map.entry(UC7ShipperSubmitBookingAmendment.class, this::sendUpdatedConfirmedBooking),
+        Map.entry(UC9ShipperCancelBookingAmendment.class, this::sendCancelBookingAmendment),
+        Map.entry(UC11ShipperCancelBookingRequestAction.class, this::sendCancelBookingRequest),
         Map.entry(
             UC13ShipperCancelConfirmedBookingAction.class,
             this::sendConfirmedBookingCancellationRequest),
-        Map.entry(AUC_Shipper_SendInvalidBookingAction.class, this::sendInvalidBookingAction),
+        Map.entry(AUCShipperSendInvalidBookingAction.class, this::sendInvalidBookingAction),
         Map.entry(ShipperGetBookingErrorScenarioAction.class, this::getBookingRequest));
   }
